code updated

+++ src/com/servlet/business/DepartmentServlet.java
@@ -8,18 +8,9 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
-
-import org.hibernate.Query;
-import org.hibernate.Session;
-import org.hibernate.Transaction;
 
 import com.app.master.Department;
 import com.app.master.DepartmentDAO;
-
-import common.HibernateUtil;
-import common.commonmethods;
-import common.dbconnection;
 
 /**
  * Servlet implementation class department_master
@@ -70,16 +61,16 @@ public class DepartmentServlet extends HttpServlet {
 			department.setName(newdepartment);
 			departmentDAO.add(department);
 			
-			
 		}else if("delete".equalsIgnoreCase(btnclick)){
 			
 			String code = request.getParameter("txtdepname").trim();
-			Department department = departmentDAO.findById(Integer.parseInt(code));
-			departmentDAO.delete(department);
+			Department department = departmentDAO.findByCode(Integer.parseInt(code));
+			if(department != null){
+				departmentDAO.delete(department);
+			}
 		}
 		
 		// ------ getting department list ---------------
-		
 		List<Department> departmentlist= new ArrayList<Department>();
 		departmentlist = departmentDAO.getList();
